refactor(canvas): share eligible node selection
