+++ extensions/xai/tsconfig.json
@@ -62,6 +62,9 @@
       "openclaw/plugin-sdk/file-access-runtime": [
         "../../packages/plugin-sdk/dist/src/plugin-sdk/file-access-runtime.d.ts"
       ],
+      "openclaw/plugin-sdk/node-selection-runtime": [
+        "../../packages/plugin-sdk/dist/src/plugin-sdk/node-selection-runtime.d.ts"
+      ],
       "openclaw/plugin-sdk/heartbeat-runtime": [
         "../../packages/plugin-sdk/dist/src/plugin-sdk/heartbeat-runtime.d.ts"
       ],
