+++ packages/plugin-sdk/src/node-selection-runtime.ts
@@ -1,0 +1,3 @@
+// Public package facade for shared node-selection policy.
+
+export * from "../../../src/plugin-sdk/node-selection-runtime.js";
