+++ src/agents/tools/nodes-utils.ts
@@ -4,7 +4,6 @@
  * Loads paired nodes from Gateway and resolves requested/default nodes with legacy pair-list fallback.
  */
 import { normalizeOptionalLowercaseString } from "@openclaw/normalization-core/string-coerce";
-import { formatErrorMessage } from "../../infra/errors.js";
 import { parseNodeList, parsePairingList } from "../../shared/node-list-parse.js";
 import type { NodeListNode } from "../../shared/node-list-types.js";
 import { resolveNodeFromNodeList, resolveNodeIdFromNodeList } from "../../shared/node-resolve.js";
@@ -171,79 +170,6 @@ export function resolveNodeIdFromList(
     allowCompactDisplayName: options.allowCompactDisplayName,
     pickDefaultNode,
   });
-}
-
-/** Tool-supplied error wording for {@link resolveEligibleNodeFromList}. */
-export type EligibleNodeMessages = {
-  /** Explicit exact-id match that is not eligible; `eligibleIds` is the sorted list (or "none"). */
-  ineligibleExact: (query: string, eligibleIds: string) => string;
-  /** Display-name/query resolution among eligible nodes failed (unknown/ambiguous). */
-  nameResolveFailed: (reason: string, eligibleIds: string) => string;
-  /** No eligible node exists. */
-  noneEligible: () => string;
-  /** Several eligible nodes and no query to disambiguate. */
-  multipleEligible: (eligible: NodeListNode[]) => string;
-};
-
-function formatNodeIdList(nodes: NodeListNode[]): string {
-  return nodes.length > 0
-    ? nodes
-        .map((node) => node.nodeId)
-        .toSorted()
-        .join(", ")
-    : "none";
-}
-
-/**
- * Resolves a capability-gated node from the FULL node list, keeping control off
- * the wrong device. An exact node-id match (case-sensitive, then -insensitive to
- * mirror display-name matching) is checked against every node first, so an
- * ineligible id can never fall through to an eligible node that merely shares its
- * display name. Display-name/query resolution runs only among eligible nodes and
- * rejects ambiguity. Any tool that filters nodes by capability must resolve
- * through here rather than handing a pre-filtered list to {@link resolveNodeIdFromList}.
- */
-export function resolveEligibleNodeFromList(
-  nodes: NodeListNode[],
-  query: string | undefined,
-  isEligible: (node: NodeListNode) => boolean,
-  messages: EligibleNodeMessages,
-): NodeListNode {
-  const eligible = nodes.filter(isEligible);
-  const trimmed = query?.trim();
-  if (trimmed) {
-    const eligibleIds = formatNodeIdList(eligible);
-    const lowerTrimmed = trimmed.toLowerCase();
-    const exactNode =
-      nodes.find((node) => node.nodeId === trimmed) ??
-      nodes.find((node) => node.nodeId.toLowerCase() === lowerTrimmed);
-    if (exactNode) {
-      if (!isEligible(exactNode)) {
-        throw new Error(messages.ineligibleExact(trimmed, eligibleIds));
-      }
-      return exactNode;
-    }
-    try {
-      const nodeId = resolveNodeIdFromList(eligible, trimmed, false);
-      const match = eligible.find((node) => node.nodeId === nodeId);
-      if (match) {
-        return match;
-      }
-    } catch (err) {
-      throw new Error(messages.nameResolveFailed(formatErrorMessage(err), eligibleIds), {
-        cause: err,
-      });
-    }
-    throw new Error(`node not found: ${trimmed}`);
-  }
-  const only = eligible.length === 1 ? eligible.at(0) : undefined;
-  if (only) {
-    return only;
-  }
-  if (eligible.length === 0) {
-    throw new Error(messages.noneEligible());
-  }
-  throw new Error(messages.multipleEligible(eligible));
 }
 
 /** Loads nodes from the Gateway and resolves the requested or default node id. */
