+++ src/plugin-sdk/node-selection-runtime.ts
@@ -1,0 +1,4 @@
+// Shared node-selection policy for bundled plugin runtime code.
+
+export type { EligibleNodeMessages } from "../shared/node-resolve.js";
+export { resolveEligibleNodeFromList } from "../shared/node-resolve.js";
